Make issuer matcher a configurable option
Main reason for this is to make the issuer something that can be overridden at
middleware configuration time.

Unfortunately, this got a little bit away from me as I refactored to make this
make more sense to me. I may have gone too far, so let me know.

The bulk of the refactoring here should have no impact on external
functionality, with the exception of how the system handles expired certificates
in its cache. Previously, we would check to see if certificates were expired
through interrogation, which meant that the certificate wrapper never had an
opportunity to react. An expired certificate, therefore, would trigger failed
authorizations until the cache expired and a new certificate was fetched.

Now, certificate expiration is handled internally to the certificate, meaning
that it is able to detect that the cert has expired before returning it, and
attempt to fetch a new, unexpired one. This simplifies the middleware class
significantly and I think that the Certificate class (previously PublicKey) is
not made much more complex in the process.

I've also made a few things constants that should have been constants, and made
some methods that I don't think we want other code relying on private (it is
generally internal functionality, like the steps required to verify whether a
token is valid).