Support wildcard resources for tokens
Adds one method, `globally_authorized?` and passes through to it when
`authorized?` is called with a scope explicitly provided.

For a few reasons, I don't think it makes sense to ever ask if the
token is globally authorized without also naming a scope, so we don't
allow that currently. Obviously, that can be changed, but I'd like to
hear more about that use case.

I currently treat `0` as a wildcard if the preferred wildcard token,
`*`, is not present. That's because it will avoid requiring a deploy
to id before this starts working. Happy to pull that out, as I would
expect we want to do so at some point anyway.